requirements.txt
-i https://pypi.org/simple
amqp==5.3.1; python_version >= '3.6'
asgiref==3.11.1; python_version >= '3.9'
bidict==0.23.1; python_version >= '3.8'
billiard==4.2.4; python_version >= '3.7'
blinker==1.9.0; python_version >= '3.9'
brotli==1.2.0
celery==5.6.3; python_version >= '3.9'
certifi==2026.2.25; python_version >= '3.7'
cffi==2.0.0; python_full_version >= '3.9' and platform_python_implementation != 'PyPy'
charset-normalizer==3.4.7; python_version >= '3.7'
click==8.3.2; python_version >= '3.10'
click-didyoumean==0.3.1; python_full_version >= '3.6.2'
click-plugins==1.1.1.2
click-repl==0.3.0; python_version >= '3.6'
configargparse==1.7.5; python_version >= '3.6'
cryptography==46.0.7; python_version >= '3.8' and python_full_version not in '3.9.0, 3.9.1'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django==6.0.4; python_version >= '3.12'
django-cors-headers==4.9.0; python_version >= '3.9'
django-filter==25.2; python_version >= '3.10'
django-tenant-schemas==2.0.0; python_version >= '3.8'
djangorestframework==3.17.1; python_version >= '3.10'
djangorestframework-simplejwt==5.5.1; python_version >= '3.9'
djoser==2.3.3; python_version >= '3.9' and python_version < '4.0'
flask==3.1.3; python_version >= '3.9'
flask-cors==6.0.2; python_version >= '3.9' and python_version < '4.0'
flask-login==0.6.3; python_version >= '3.7'
gevent==25.9.1; python_version >= '3.9'
geventhttpclient==2.3.9; python_version >= '3.9'
greenlet==3.4.0; platform_python_implementation == 'CPython'
gunicorn==25.3.0; python_version >= '3.10'
h11==0.16.0; python_version >= '3.8'
idna==3.11; python_version >= '3.8'
iniconfig==2.3.0; python_version >= '3.10'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.6; python_version >= '3.7'
kombu==5.6.2; python_version >= '3.9'
locust==2.43.4; python_version >= '3.10'
markupsafe==3.0.3; python_version >= '3.9'
msgpack==1.1.2; python_version >= '3.9'
oauthlib==3.3.1; python_version >= '3.8'
ordered-set==4.1.0; python_version >= '3.7'
packaging==26.0; python_version >= '3.8'
paypalrestsdk==1.13.3
pillow==12.2.0; python_version >= '3.10'
pluggy==1.6.0; python_version >= '3.9'
prompt-toolkit==3.0.52; python_version >= '3.8'
psutil==7.2.2; python_version >= '3.6'
psycopg2-binary==2.9.11; python_version >= '3.9'
pycparser==3.0; implementation_name != 'PyPy'
pygments==2.20.0; python_version >= '3.9'
pyjwt[crypto]==2.12.1; python_version >= '3.9'
pyopenssl==26.0.0; python_version >= '3.8'
pytest==9.0.3; python_version >= '3.10'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-engineio==4.13.1; python_version >= '3.8'
python-socketio[client]==5.16.1; python_version >= '3.8'
python3-openid==3.2.0
pyzmq==27.1.0; python_version >= '3.8'
redis==7.4.0; python_version >= '3.10'
reportlab==4.4.10; python_version >= '3.9' and python_version < '4'
requests==2.33.1; python_version >= '3.10'
requests-oauthlib==2.0.0; python_version >= '3.4'
simple-websocket==1.1.0; python_version >= '3.6'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
social-auth-app-django==5.7.0; python_version >= '3.10'
social-auth-core==4.8.5; python_version >= '3.10'
sqlparse==0.5.5; python_version >= '3.8'
stripe==15.0.1; python_version >= '3.9'
typing-extensions==4.15.0; python_version >= '3.9'
tzdata==2026.1; python_version >= '2'
tzlocal==5.3.1; python_version >= '3.9'
urllib3==2.6.3; python_version >= '3.9'
vine==5.1.0; python_version >= '3.6'
wcwidth==0.6.0; python_version >= '3.8'
websocket-client==1.9.0
werkzeug==3.1.8; python_version >= '3.9'
whitenoise==6.12.0; python_version >= '3.10'
wsproto==1.3.2; python_version >= '3.10'
zope.event==6.1; python_version >= '3.10'
zope.interface==8.3; python_version >= '3.10'
